+++ internal/telegram/pipeline.go
@@ -100,7 +100,7 @@ func (bc *BotController) processBridgeEventsAsyncWithThread(chat *telebot.Chat, 
 	if len(userID) > 0 {
 		uid = userID[0]
 	}
-	return bridgeOutcome(bc.ensurePipeline().ProcessBridgeEvents(chat.ID, threadID, messageID, ch, progress, userText, nil, uid))
+	return bridgeOutcome(bc.ensurePipeline().ProcessBridgeEvents(chat.ID, threadID, messageID, ch, progress, userText, nil, uid, nil, nil))
 }
 
 func (bc *BotController) invalidateMemoryDirs(chatID int64, threadID int, userID int64, cwd string) {